Finishing metrics.py refactor
