+++ aisdc/attacks/worst_case_attack.py
@@ -17,7 +17,8 @@ from sklearn.ensemble import RandomForestClassifier
 from sklearn.metrics import confusion_matrix
 from sklearn.model_selection import train_test_split
 
-from aisdc.attacks import metrics, report
+from aisdc import metrics
+from aisdc.attacks import report
 from aisdc.attacks.attack import Attack
 from aisdc.attacks.dataset import Data
 
@@ -227,9 +228,10 @@ class WorstCaseAttack(Attack):
                 **self.args.mia_attack_model_hyp
             )
             attack_classifier.fit(mi_train_x, mi_train_y)
+            y_pred_proba, y_test = metrics.get_probabilities(attack_classifier, mi_test_x, mi_test_y,permute_rows=True)
 
             mia_metrics.append(
-                metrics.get_metrics(attack_classifier, mi_test_x, mi_test_y)
+                metrics.get_metrics(y_pred_proba, y_test)
             )
 
             if self.args.include_model_correct_feature and train_correct is not None:
