+++ aisdc/metrics.py
@@ -255,6 +255,30 @@ def get_probabilities(  # pylint: disable=too-many-locals
     y_test: np.ndarray = [],
     permute_rows: bool = None
 ):
+  
+    """
+    Given a prediction model and a dataset, calculate the predictions of the model for
+    each data sample in probability format
+    Parameters
+    ----------
+    clf: sklearn.Model
+        trained model
+    X_test: np.ndarray
+        test data matrix
+    y_test: np.ndarray
+        test data labels
+    permute_rows: boolean
+        a flag to indicate whether rows should be permuted
+    Returns
+    -------
+    y_pred_proba: a list of probabilities for each sample in the dataset
+    
+    Notes
+    -----
+    If permute_rows is set to true, y_test must also be supplied.
+    The function will then return both the predicted probabilities and corresponding y_test
+    """
+
     if permute_rows == True and len(y_test) == 0:
         raise ValueError ("If permute_rows is set to True, y_test must be supplied")
 
@@ -273,8 +297,8 @@ def get_probabilities(  # pylint: disable=too-many-locals
     return y_pred_proba
 
 def get_metrics(  # pylint: disable=too-many-locals
-    y_test: np.ndarray,
     y_pred_proba: np.ndarray,
+    y_test: np.ndarray
 ):
 
     invalid_format_error_message = "y_pred must be an array of shape [x,2] with elements of type float"
@@ -291,8 +315,6 @@ def get_metrics(  # pylint: disable=too-many-locals
     except ValueError:
         print(invalid_format_error_message)
 
-    print("New Metrics")
-    print(y_test.shape)
     """
     Calculate metrics, including attacker advantage for MIA binary.
     Implemented as Definition 4 on https://arxiv.org/pdf/1709.01604.pdf
@@ -300,12 +322,10 @@ def get_metrics(  # pylint: disable=too-many-locals
 
     Parameters
     ----------
-    clf: sklearn.Model
-        trained model
-    X_test: np.ndarray
-        test data matrix
     y_test: np.ndarray
         test data labels
+    y_pred_proba: np.ndarray of shape [x,2] and type float
+        predicted probabilities
 
     Returns
     -------
@@ -331,7 +351,6 @@ def get_metrics(  # pylint: disable=too-many-locals
     """
     metrics = {}
 
-    print(y_pred_proba.shape)
     y_pred = np.argmax(y_pred_proba,axis=1)
     y_pred_proba = y_pred_proba[:, 1]
 
@@ -376,10 +395,6 @@ def get_metrics(  # pylint: disable=too-many-locals
     metrics["FDIF01"] = fdif
     metrics["PDIF01"] = -pdif  # use -log(p) so answer is positive
 
-    print("In New Metrics")
-    print(y_test.shape)
-    print(y_pred_proba.shape)
-
     fmax, fmin, fdif, pdif = min_max_disc(y_test, y_pred_proba, x_prop=0.2)
     metrics["FMAX02"] = fmax
     metrics["FMIN02"] = fmin
@@ -411,141 +426,3 @@ def get_metrics(  # pylint: disable=too-many-locals
     metrics["n_neg_test_examples"] = (1 - y_test).sum()
 
     return metrics
-
-import pandas as pd
-import numpy as np
-from sklearn.model_selection import train_test_split
-from sklearn.ensemble import RandomForestClassifier
-from sklearn import metrics
-
-def format_column(df, name):
-    df[name] = pd.Categorical(df[name])
-    df[name+'_code'] = df[name].cat.codes
-    return df
-
-def binary_classification():
-    df = pd.read_csv("Customer_Churn.csv")
-
-    df = format_column(df, 'gender')
-    df = format_column(df, 'SeniorCitizen')
-    df = format_column(df, 'PhoneService')
-    df = format_column(df, 'MultipleLines')
-    df = format_column(df, 'PaymentMethod')
-    df = format_column(df, 'Partner')
-    df = format_column(df, 'Dependents')
-    df = format_column(df, 'PaperlessBilling')
-    df = format_column(df, 'OnlineSecurity')
-    df = format_column(df, 'TechSupport')
-    df = format_column(df, 'InternetService')
-    df = format_column(df, 'Contract')
-    df = format_column(df, 'StreamingTV')
-    df = format_column(df, 'DeviceProtection')
-    df = format_column(df, 'StreamingMovies')
-    df = format_column(df, 'OnlineBackup')
-    df = format_column(df, 'Churn')
-
-    features = ['gender_code', 'SeniorCitizen_code', 'PhoneService_code', 'MultipleLines_code', 
-        'InternetService_code', 'Partner_code', 'Dependents_code', 'PaymentMethod_code', 'PaperlessBilling_code','Contract_code', 'StreamingMovies_code',
-        'StreamingTV_code', 'TechSupport_code', 'DeviceProtection_code', 'OnlineBackup_code',
-        'OnlineSecurity_code', 'Dependents_code', 'Partner_code','tenure', 'MonthlyCharges']
-
-    X = np.array(df[features])
-    y = np.array(df['Churn_code'])
-
-    return X, y
-
-def multiclass_classification():
-    attributes = ["sepal_length", "sepal_width", "petal_length", "petal_width", "species"]
-    dataset = pd.read_csv('iris.data', names = attributes)
-    dataset.columns = attributes
-
-    print(dataset.head())
-    print(dataset['species'].value_counts())
-
-    y = dataset['species']
-    X = dataset.drop(columns=['species'])
-
-    return X, y
-
-if __name__ == "__main__":
-    print("Hello world!")
-    
-    X, y = binary_classification()
-
-    x_train, x_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
-
-    rf = RandomForestClassifier()
-    rf.fit(x_train, y_train)
-    y_pred = rf.predict(x_test)
-
-    # print(metrics.classification_report(y_test, y_pred))
-
-    # tn, fp, fn, tp = confusion_matrix(y_test, y_pred).ravel()
-    # tpr = round(float(tp / (tp + fn)), 8)
-    # fpr = round(float(fp / (fp + tn)), 8)
-    # tnr = round(float(tn / (tn + fp)), 8)
-    # fnr = round(float(fn / (tp + fn)), 8)
-
-    # print("True positive: "+str(tpr))
-    # print("False positive: "+str(fpr))
-    # print("True negative: "+str(tnr))
-    # print("False negative: "+str(fnr))
-
-    # y_pred = rf.predict_proba(x_test)
-
-    print("SHAPES")
-    print(y_pred.shape)
-    print(y_test.shape)
-
-    y_pred_proba = get_probabilities(rf, x_test, permute_rows=False)
-
-    print(y_pred_proba.shape)
-
-    print(new_get_metrics(y_test, y_pred_proba))
-
-    y_pred = rf.predict_proba(x_train)
-    y_label = [1]*len(y_pred)
-    print(len(y_label))
-
-    y_pred_test = rf.predict_proba(x_test)
-    y_label_y = [0]*len(y_pred_test)
-    
-    for y in y_label_y:
-        y_label.append(y)
-    print(len(y_label))
-    print("End of ylabel")
-
-    y_pred = np.append(y_pred, y_pred_test, axis=0)
-
-    print(len(y_pred))
-    print(len(y_label))
-
-    print(y_pred.shape)
-
-    attack_rf = RandomForestClassifier()
-    x_train_attack, x_test_attack, y_train_attack, y_test_attack = train_test_split(np.array(y_pred), np.array(y_label), test_size=0.2, random_state=42)
-
-    attack_rf.fit(x_train_attack,y_train_attack)
-
-    permute_rows = False
-    if permute_rows:
-        N, _ = x_test.shape
-        order = np.random.RandomState(  # pylint: disable = no-member
-            seed=10
-        ).permutation(N)
-        x_train_attack = x_train_attack[order, :]
-        y_test_attack = y_test_attack[order]
-    y_pred = attack_rf.predict_proba(x_test_attack)
-
-    before = get_metrics(attack_rf, x_test_attack, y_test_attack, permute_rows=False)
-
-    y_pred_proba = get_probabilities(attack_rf, x_test_attack, permute_rows=False)
-    after = new_get_metrics(y_test_attack, y_pred_proba)
-
-    for key in before.keys():
-        b = before[key]
-        a = after[key]
-
-        print(key)
-        print(a == b)
-        print()
